userIntger

+++ src/main/java/com/demo/demo/repository/RoleRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> { // Long au lieu de Integer
 
+    // ✅ Le nom du champ dans l'entity est "rolename"
     Optional<Role> findByRolename(RoleName rolename);
-
 
 }
